refactor: delete the OS hook now the harness runs its tools
