+++ apps/tui/Cargo.toml
@@ -15,7 +15,6 @@ name = "crabtalk"
 path = "src/bin/main.rs"
 
 [dependencies]
-hooks = { workspace = true, default-features = false, features = ["os"] }
 client.workspace = true
 transport.workspace = true
 wcore.workspace = true
