+++ crates/hooks/src/lib.rs
@@ -17,8 +17,6 @@ use wcore::{AgentConfig, AgentEvent, ToolDispatch, ToolFuture};
 pub mod mcp;
 #[cfg(feature = "memory")]
 pub mod memory;
-#[cfg(feature = "os")]
-pub mod os;
 #[cfg(feature = "skill")]
 pub mod skill;
 #[cfg(feature = "memory")]
@@ -28,8 +26,6 @@ mod utils;
 pub use mcp::McpHook;
 #[cfg(feature = "memory")]
 pub use memory::{DEFAULT_SOUL, Memory, MemoryHook};
-#[cfg(feature = "os")]
-pub use os::OsHook;
 #[cfg(feature = "skill")]
 pub use skill::handler::SkillHook;
 #[cfg(feature = "memory")]
